+++ crates/miden-objects/src/account/file.rs
@@ -15,7 +15,7 @@ use super::super::utils::serde::{
     DeserializationError,
     Serializable,
 };
-use super::{Account, AuthSecretKey, Word};
+use super::{Account, AuthSecretKey};
 
 const MAGIC: &str = "acct";
 
@@ -33,21 +33,12 @@ const MAGIC: &str = "acct";
 #[derive(Debug, Clone)]
 pub struct AccountFile {
     pub account: Account,
-    pub account_seed: Option<Word>,
     pub auth_secret_keys: Vec<AuthSecretKey>,
 }
 
 impl AccountFile {
-    pub fn new(
-        account: Account,
-        account_seed: Option<Word>,
-        auth_keys: Vec<AuthSecretKey>,
-    ) -> Self {
-        Self {
-            account,
-            account_seed,
-            auth_secret_keys: auth_keys,
-        }
+    pub fn new(account: Account, auth_keys: Vec<AuthSecretKey>) -> Self {
+        Self { account, auth_secret_keys: auth_keys }
     }
 }
 
@@ -76,14 +67,9 @@ impl AccountFile {
 impl Serializable for AccountFile {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
         target.write_bytes(MAGIC.as_bytes());
-        let AccountFile {
-            account,
-            account_seed,
-            auth_secret_keys: auth,
-        } = self;
+        let AccountFile { account, auth_secret_keys: auth } = self;
 
         account.write_into(target);
-        account_seed.write_into(target);
         auth.write_into(target);
     }
 }
@@ -97,10 +83,9 @@ impl Deserializable for AccountFile {
             )));
         }
         let account = Account::read_from(source)?;
-        let account_seed = <Option<Word>>::read_from(source)?;
         let auth_secret_keys = <Vec<AuthSecretKey>>::read_from(source)?;
 
-        Ok(Self::new(account, account_seed, auth_secret_keys))
+        Ok(Self::new(account, auth_secret_keys))
     }
 
     fn read_from_bytes(bytes: &[u8]) -> Result<Self, DeserializationError> {
@@ -120,7 +105,7 @@ mod tests {
     use tempfile::tempdir;
 
     use super::AccountFile;
-    use crate::account::{Account, AccountCode, AccountId, AuthSecretKey, Felt, Word, storage};
+    use crate::account::{Account, AccountCode, AccountId, AuthSecretKey, Felt, storage};
     use crate::asset::AssetVault;
     use crate::testing::account_id::ACCOUNT_ID_REGULAR_PUBLIC_ACCOUNT_IMMUTABLE_CODE;
 
@@ -133,11 +118,10 @@ mod tests {
         let storage = AccountStorage::new(vec![]).unwrap();
         let nonce = Felt::new(1);
         let account = Account::new_existing(id, vault, storage, code, nonce);
-        let account_seed = Some(Word::empty());
         let auth_secret_key = AuthSecretKey::RpoFalcon512(SecretKey::new());
         let auth_secret_key_2 = AuthSecretKey::RpoFalcon512(SecretKey::new());
 
-        AccountFile::new(account, account_seed, vec![auth_secret_key, auth_secret_key_2])
+        AccountFile::new(account, vec![auth_secret_key, auth_secret_key_2])
     }
 
     #[test]
@@ -146,7 +130,6 @@ mod tests {
         let serialized = account_file.to_bytes();
         let deserialized = AccountFile::read_from_bytes(&serialized).unwrap();
         assert_eq!(deserialized.account, account_file.account);
-        assert_eq!(deserialized.account_seed, account_file.account_seed);
         assert_eq!(
             deserialized.auth_secret_keys.to_bytes(),
             account_file.auth_secret_keys.to_bytes()
@@ -164,7 +147,6 @@ mod tests {
         let deserialized = AccountFile::read(filepath.as_path()).unwrap();
 
         assert_eq!(deserialized.account, account_file.account);
-        assert_eq!(deserialized.account_seed, account_file.account_seed);
         assert_eq!(
             deserialized.auth_secret_keys.to_bytes(),
             account_file.auth_secret_keys.to_bytes()